requirements.txt
aws-cdk-lib==2.15.0
constructs>=10.0.0,<11.0.0
aws_lambda_powertools
boto3
pylint
pdoc3
yapf
requests
bs4
pytz